final changes

+++ system/Employee.cs
@@ -22,7 +22,7 @@ namespace system
         {
             if (auth && trip.hasEmptySeats())
             {
-                int ticketId = Int32.Parse(trip.date.ToString("yyyyMMddHHmm") + trip.tickets.Count.ToString() + trip.id.ToString());
+                long ticketId = long.Parse(trip.date.ToString("yyyyMMddHHmm") + trip.tickets.Count.ToString() + trip.id.ToString());
                 var ticket = new OfflineTicket(this, trip, ticketId);
                 tickets.Add(ticket);
                 trip.addTicket(ticket);
